requirements.txt
pyTelegramBotAPI
selenium
webdriver-manager